==> analysis/01_imputation-prep-unitlevel-estimation-highres.R <==
################################################################################
################## POVERTY MAPPING FOR ADMIN 3 AND ADMIN 4 #####################
################################################################################
pacman::p_load("haven", "sf", "data.table", "dplyr", "ggplot2", "nlme", "RStata",
               "exactextractr", "raster", "povmap", "viridis", "glmmLasso",
               "lme4", "gridExtra", "MASS", "furrr", "purrr", "survey")

sf::sf_use_s2(FALSE)

### read in the datasets

grid_dt <- readRDS("data-raw/ind_list.RDS")

grid_dt <- grid_dt$admin4_poppoly ### the geospatial grid data

pop_dt <- readRDS("data-clean/worldpop.RDS")

grid_dt$admin4Pcode <- as.integer(grid_dt$admin4Pcode)
pop_dt$admin4Pcode <- as.integer(pop_dt$admin4Pcode)

grid_dt <- merge(grid_dt,
                 pop_dt[,c("admin4Pcode",
                           "bdi_ppp_2020_UNadj_constrained")] %>%
                   st_drop_geometry())

geosurvey_dt <- haven::read_dta("data-raw/hh_poverty.dta") ### household survey


geosurvey_dt <-
  geosurvey_dt %>%
  rename(welfare = "rpc_tot_cons")

geosurvey_dt %>%
  mutate(poor = ifelse(welfare < pline_int_215, 1, 0)) %>%
  summarise(weighted.mean(x = poor,
                          w = weight_adj * hh_size,
                          na.rm = TRUE))


geocodes_dt <-
  haven::read_spss("data-raw/Coordonnées GPS des ménages enquêtés.sav")

geocodes_dt <- as.data.table(geocodes_dt)
geosurvey_dt <- as.data.table(geosurvey_dt)

setnames(geocodes_dt, c("IDZD", "POIDS"), c("enum_area", "weight"))

geosurvey_dt[, hhno := as.integer(substr(hhid,
                                         nchar(as.character(hhid)) - 1,
                                         nchar(as.character(hhid))))]

## create a household ID from the geocode data
geocodes_dt[, hhid := paste0(enum_area,
                             sprintf("%02d", ID03))]

geocodes_dt[, hhid := as.integer(hhid)]

geosurvey_dt <- geocodes_dt[geosurvey_dt, on = "hhid"][, c("i.enum_area",
                                                           "i.weight") := NULL ]


### read in the new admin shapefiles
shp_dt <- sf::st_read(dsn = "data-raw/diva-gis",
                      layer = "BDI_adm4")


### relabel shapefile area names
shp_dt <-
  shp_dt %>%
  rename(admin0Pcod = "ISO",
         admin0Name = "NAME_0",
         admin1Pcod = "ID_1",
         admin1Name = "NAME_1",
         admin2Pcod = "ID_2",
         admin2Name = "NAME_2",
         admin3Pcod = "ID_3",
         admin3Name = "NAME_3",
         admin4Pcod = "ID_4",
         admin4Name = "NAME_4")

### check the merge between survey, grid and new shapefile
cols_to_remove <- colnames(grid_dt)[grep("^admin.*(?:Name|Pcode|Pcod)$", colnames(grid_dt))]

grid_dt <- grid_dt[, -cols_to_remove, with = F] ##drop the old admin column names

gridpoly_dt <-
  grid_dt %>%
  mutate(grid_ID = 1:nrow(.))

gridpoly_dt <-
  gridpoly_dt[, c("grid_ID", "geometry")] %>%
  st_as_sf(crs = 32735) %>%
  st_transform(crs = 4326)

grid_dt <-
  grid_dt %>%
  mutate(grid_ID = 1:nrow(.)) %>%
  as.data.frame() %>%
  st_as_sf(crs = 32735) %>%
  st_transform(crs = 4326) %>%
  st_centroid() %>%
  st_join(shp_dt)

missing_admin <-
st_nearest_feature(x = grid_dt[is.na(grid_dt$admin4Pcod),],
                   y = shp_dt)

missing_dt <- shp_dt[missing_admin,]

add_dt <-
  grid_dt[is.na(grid_dt$admin4Pcod), colnames(grid_dt)[!colnames(grid_dt) %in% colnames(shp_dt)]] %>%
  cbind(missing_dt %>%
          st_drop_geometry())

### fix the variable names in add_dt to match grid_dt
colnames(add_dt) <- sub("^X(\\d{4})", "\\1", colnames(add_dt))


grid_dt <-
  grid_dt[!is.na(grid_dt$admin4Pcod),] %>%
  rbind(add_dt)

geosurvey_dt <-
  geosurvey_dt[, c("hhid", "hh_size", "weight", "weight_adj",
                   "welfare", "pline_abs_lb_ae",
                   "pline_int_215", "SI15", "SI16")] %>%
  st_as_sf(crs = 4326,
           coords = c("SI16", "SI15")) %>%
  st_join(shp_dt)

### create a geospatial pseudo-welfare aggregate variable at the grid level using NTL
hhgeo_plot <-
  ggplot() +
  geom_sf(data = shp_dt, fill = "white", color = "black") +
  geom_sf(data = geosurvey_dt %>%
            as.data.frame() %>%
            st_as_sf(),
          color = "red") +
  labs(title = "Household locations within Burundi",
       x = "Longitude",
       y = "Latitude") +
  theme_minimal()

ggsave(plot = hhgeo_plot,
       "figures/hhlocations_plot.png")

### create some variables for growth in ntl
grid_dt <- as.data.table(grid_dt)

ntl_all_cols <- c("ntl_all_2015", "ntl_all_2016", "ntl_all_2017",
                  "ntl_all_2018", "ntl_all_2019", "ntl_all_2020")

grid_dt[, ntl_all_cols] <- grid_dt[, ntl_all_cols, with = F] + 1

ntl_built_up_cols <- c("ntl_built_up_2015", "ntl_built_up_2016", "ntl_built_up_2017",
                       "ntl_built_up_2018", "ntl_built_up_2019", "ntl_built_up_2020")

grid_dt[, ntl_built_up_cols] <- grid_dt[, ntl_built_up_cols, with = F] + 1

ntl_small_stl_cols <- c("ntl_small_stl_2015", "ntl_small_stl_2016", "ntl_small_stl_2017",
                        "ntl_small_stl_2018", "ntl_small_stl_2019", "ntl_small_stl_2020")

grid_dt[, ntl_small_stl_cols] <- grid_dt[, ntl_small_stl_cols, with = F] + 1

ntl_hamlet_cols <- c("ntl_hamlet_2015", "ntl_hamlet_2016", "ntl_hamlet_2017",
                     "ntl_hamlet_2018", "ntl_hamlet_2019", "ntl_hamlet_2020")

grid_dt[, ntl_hamlet_cols] <- grid_dt[, ntl_hamlet_cols, with = F] + 1

ntl_cols <- c(ntl_all_cols, ntl_built_up_cols, ntl_small_stl_cols, ntl_hamlet_cols)

grid_dt[, ntl_cols] <- log(grid_dt[, ntl_cols, with = F])

grid_dt[, ntl_all_growth := ntl_all_2020 - ntl_all_2015]
grid_dt[, ntl_built_up_growth := ntl_built_up_2020 - ntl_built_up_2015]
grid_dt[, ntl_small_stl_growth := ntl_small_stl_2020 - ntl_small_stl_2015]
grid_dt[, ntl_hamlet_growth := ntl_hamlet_2020 - ntl_hamlet_2015]


### growth in population density
grid_dt[, log_pop_density_growth := log(pop_density_2020 + 1) -
          log(pop_density_2015 + 1)]

### growth in each of the different landcover class types
grid_dt[, water_growth := `2020_Water` - `2017_Water`]
grid_dt[, trees_growth := `2020_Trees` - `2017_Trees`]
grid_dt[, floodedveg_growth := `2020_Flooded.vegetation` - `2017_Flooded.vegetation`]
grid_dt[, crops_growth := `2020_Crops` - `2017_Crops`]
grid_dt[, builtarea_growth := `2020_Built.Area` - `2017_Built.Area`]
grid_dt[, bareground_growth := `2020_Bare_ground` - `2017_Bare_ground`]
grid_dt[, snowice_growth := `2020_Snow.Ice` - `2017_Snow.Ice`]
grid_dt[, clouds_growth := `2020_Clouds` - `2017_Clouds`]
grid_dt[, rangeland_growth := `2020_Rangeland` - `2017_Rangeland`]



##### compute the regional welfare totals for each province
add_dt <-
geosurvey_dt %>%
  st_drop_geometry() %>%
  group_by(admin1Pcod) %>%
  summarise(welfare = sum(welfare * weight_adj * hh_size, na.rm = TRUE)) %>%
  merge(grid_dt[, c("grid_ID", "admin1Pcod", "ntl_all_2015",
                    "ntl_all_2016", "ntl_all_2017", "ntl_all_2018",
                    "ntl_all_2019", "ntl_all_2020",
                    "bdi_ppp_2020_UNadj_constrained")] %>%
          st_drop_geometry()) %>%
  group_by(admin1Pcod) %>%
  mutate(ntl2015_welf_int = log(1 + (((ntl_all_2015 / sum(ntl_all_2015)) * welfare))/bdi_ppp_2020_UNadj_constrained)) %>%
  mutate(ntl2016_welf_int = log(1 + (((ntl_all_2016 / sum(ntl_all_2016)) * welfare))/bdi_ppp_2020_UNadj_constrained)) %>%
  mutate(ntl2017_welf_int = log(1 + (((ntl_all_2017 / sum(ntl_all_2017)) * welfare))/bdi_ppp_2020_UNadj_constrained)) %>%
  mutate(ntl2018_welf_int = log(1 + (((ntl_all_2018 / sum(ntl_all_2018)) * welfare))/bdi_ppp_2020_UNadj_constrained)) %>%
  mutate(ntl2019_welf_int = log(1 + (((ntl_all_2019 / sum(ntl_all_2019)) * welfare))/bdi_ppp_2020_UNadj_constrained)) %>%
  mutate(ntl2020_welf_int = log(1 + (((ntl_all_2020 / sum(ntl_all_2020)) * welfare))/bdi_ppp_2020_UNadj_constrained))

add_dt$ntl2016_welf_int[is.na(add_dt$ntl2016_welf_int)] <- 0

grid_dt <-
  grid_dt %>%
  merge(add_dt[, c("grid_ID", "ntl2015_welf_int", "ntl2016_welf_int",
                   "ntl2017_welf_int", "ntl2018_welf_int",
                   "ntl2019_welf_int", "ntl2020_welf_int")])

##### target area variables at admin4, admin3 and admin2
### first make sure all variable names are properly created
colnames(grid_dt) <- gsub("^(\\d)", "y\\1", colnames(grid_dt))
colnames(grid_dt) <- gsub("\\.", "_", colnames(grid_dt))

### rename variable names that are over 32 characters long
invalid_names <- colnames(grid_dt)[nchar(colnames(grid_dt)) >= 32]

valid_names <- gsub(pattern = "_MS_MEAN_v01",
                    replacement = "",
                    invalid_names)

setnames(grid_dt, invalid_names, valid_names)

#### select the set of candidate variables
drop_vars <- c("grid_ID", "area", "poly_area", "ID_0", "admin0Pcod",
               "admin0Name", "admin1Pcod", "admin1Name", "admin2Pcod",
               "admin2Name", "admin3Name", "admin3Pcod", "admin4Name",
               "admin4Pcod", "VARNAME_4", "TYPE_4", "ENGTYPE_4",
               "geometry", "")
candidate_vars <- colnames(grid_dt)[!colnames(grid_dt) %in% drop_vars]

#### replace missing values with admin1 averages
grid_dt[, (candidate_vars) :=
          lapply(.SD, function(x){

            if (any(is.na(x))) {
              mean_val <- mean(x, na.rm = TRUE)
              replace(x, is.na(x), mean_val)
            } else {
              x
            }

          }),
        .SDcols = candidate_vars,
        by = "admin1Pcod"]

## some variables have super large values lets transform them!
columns_to_scale <- sapply(as.data.frame(grid_dt[,candidate_vars, with = F]), function(x) any(abs(x) > 100))

columns_to_scale <- columns_to_scale[columns_to_scale == TRUE & is.na(columns_to_scale) == FALSE]
columns_to_scale <- names(columns_to_scale)
#### log the building and population variables
log_vars <- columns_to_scale[grepl("density", columns_to_scale) | grepl("building", columns_to_scale)]

grid_dt[, (log_vars) := lapply(.SD,
                               function(x)(log(x + 1))),
        .SDcols = log_vars]

scale_vars <- columns_to_scale[!(columns_to_scale %in% log_vars)]
scale_vars <- scale_vars[!grepl("bdi_ppp_2020_UNadj_constrained", scale_vars)]

grid_dt[, (scale_vars) := lapply(.SD, scale, center = TRUE), .SDcols = scale_vars]

#### remove variables that are constant
constant_cols <- names(grid_dt)[sapply(grid_dt, function(x) length(unique(x)) == 1)]

grid_dt[, (constant_cols) := NULL]

candidate_vars <- candidate_vars[!(candidate_vars %in% constant_cols)]
candidate_vars <- candidate_vars[!grepl(pattern = "bdi_ppp_2020_UNadj_constrained",
                                       x = candidate_vars)]

### now finally compute the admin average variables
admin1_dt <-
  grid_dt[, lapply(.SD,
                   weighted.mean,
                   w = bdi_ppp_2020_UNadj_constrained,
                   na.rm = TRUE),
          .SDcols = candidate_vars,
          by = "admin1Pcod"]

colnames(admin1_dt)[colnames(admin1_dt) %in% candidate_vars] <-
  paste0("admin1_", colnames(admin1_dt)[colnames(admin1_dt) %in% candidate_vars])

admin2_dt <-
  grid_dt[, lapply(.SD, weighted.mean,
                   w = bdi_ppp_2020_UNadj_constrained,
                   na.rm = TRUE),
          .SDcols = candidate_vars,
          by = "admin2Pcod"]

colnames(admin2_dt)[colnames(admin2_dt) %in% candidate_vars] <-
  paste0("admin2_", colnames(admin2_dt)[colnames(admin2_dt) %in% candidate_vars])

admin3_dt <-
  grid_dt[, lapply(.SD,
                   weighted.mean,
                   w = bdi_ppp_2020_UNadj_constrained,
                   na.rm = TRUE),
          .SDcols = candidate_vars,
          by = "admin3Pcod"]

colnames(admin3_dt)[colnames(admin3_dt) %in% candidate_vars] <-
  paste0("admin3_", colnames(admin3_dt)[colnames(admin3_dt) %in% candidate_vars])

admin4_dt <-
  grid_dt[, lapply(.SD,
                   weighted.mean,
                   w = bdi_ppp_2020_UNadj_constrained,
                   na.rm = TRUE),
          .SDcols = candidate_vars,
          by = "admin4Pcod"]

colnames(admin4_dt)[colnames(admin4_dt) %in% candidate_vars] <-
  paste0("admin4_", colnames(admin4_dt)[colnames(admin4_dt) %in% candidate_vars])


grid_dt <- admin1_dt[grid_dt, on = "admin1Pcod"]
grid_dt <- admin2_dt[grid_dt, on = "admin2Pcod"]
grid_dt <- admin3_dt[grid_dt, on = "admin3Pcod"]
grid_dt <- admin4_dt[grid_dt, on = "admin4Pcod"]

cols_to_remove <- colnames(grid_dt)[grep("^admin.*(?:Name|Pcode|Pcod)$", colnames(grid_dt))]


#### merge in the geo household survey

grid_dt <-
  grid_dt %>%
  as.data.frame() %>%
  st_as_sf(crs = 4326) %>%
  st_drop_geometry() %>%
  merge(gridpoly_dt, by = "grid_ID")

drop_vars <- colnames(grid_dt)[colnames(grid_dt) %in% colnames(geosurvey_dt)]

geosurvey_dt <-
geosurvey_dt %>%
  st_join(grid_dt[,c(colnames(grid_dt)[!colnames(grid_dt) %in% colnames(geosurvey_dt)], "geometry")] %>%
            st_as_sf(crs = 4326))

geosurvey_dt <-
geosurvey_dt %>%
  mutate(lnwelfare = log(welfare + 1))

### include area dummies
### create area level dummies
add_dt <- dummify(as.character(geosurvey_dt$admin1Pcod))
colnames(add_dt) <- paste0("BDI", sprintf("%02d", as.integer(colnames(add_dt))))
geosurvey_dt <- cbind(geosurvey_dt,
                      as.data.table(add_dt))

add_dt <- dummify(as.character(grid_dt$admin1Pcod))
colnames(add_dt) <- paste0("BDI", sprintf("%02d", as.integer(colnames(add_dt))))
grid_dt <- cbind(grid_dt,
                 as.data.table(add_dt))


geosurvey_dt <- as.data.table(geosurvey_dt)
grid_dt <- as.data.table(grid_dt)
################################################################################
############## PREPARE FOR MODEL SELECTION FOR MIXED MODELS ####################
################################################################################
### perform the model selection with the lasso linear model
drop_vars <- c("grid_ID", cols_to_remove, "geometry", "TYPE_4", "ENGTYPE_4",
               "area")

candidate_vars <- colnames(grid_dt)[!colnames(grid_dt) %in% drop_vars]

geosurvey_dt[, hhweight := weight_adj * hh_size]

### create the area level dummies
### create area level dummies
geosurvey_dt <- cbind(geosurvey_dt,
                      as.data.table(dummify(geosurvey_dt$admin1Pcod)))

grid_dt <- cbind(grid_dt,
                 as.data.table(dummify(grid_dt$admin1Pcod)))

geosurvey_dt[, (candidate_vars) :=
               lapply(.SD, function(x){

                 if (any(is.na(x))) {
                   mean_val <- mean(x, na.rm = TRUE)
                   replace(x, is.na(x), mean_val)
                 } else {
                   x
                 }

               }),
             .SDcols = candidate_vars,
             by = "admin1Pcod"]

### save the data to .dta file
colnames(geosurvey_dt)[grepl("bdi_ppp_2020_UNadj_constrained", colnames(geosurvey_dt))] <-
  gsub(pattern = "bdi_ppp_2020_UNadj_constrained",
       replacement = "wpop_population",
       x = colnames(geosurvey_dt)[grepl("bdi_ppp_2020_UNadj_constrained", colnames(geosurvey_dt))])

colnames(grid_dt)[grepl("bdi_ppp_2020_UNadj_constrained", colnames(grid_dt))] <-
  gsub(pattern = "bdi_ppp_2020_UNadj_constrained",
       replacement = "wpop_population",
       x = colnames(grid_dt)[grepl("bdi_ppp_2020_UNadj_constrained", colnames(grid_dt))])

candidate_vars[grepl("bdi_ppp_2020_UNadj_constrained", candidate_vars)] <-
  gsub(pattern = "bdi_ppp_2020_UNadj_constrained",
       replacement = "wpop_population",
       x = candidate_vars[grepl("bdi_ppp_2020_UNadj_constrained", candidate_vars)])


# haven::write_dta(geosurvey_dt[, c(candidate_vars,
#                                   "lnwelfare",
#                                   "admin1Pcod",
#                                   "admin2Pcod",
#                                   "admin3Pcod",
#                                   "admin4Pcod",
#                                   "hhweight"),
#                               with = F],
#                  "data-clean/geosurvey.dta")


#### ran lasso linear model selection in stata
admin4selvars_list <- readLines("data-clean/model_selection/selected_variables_admin4.txt")
admin4selvars_list <- strsplit(admin4selvars_list, " +")[[1]]

admin3selvars_list <- readLines("data-clean/model_selection/selected_variables_admin3.txt")
admin3selvars_list <- strsplit(admin3selvars_list, " +")[[1]]

admin2selvars_list <- readLines("data-clean/model_selection/selected_variables_admin2.txt")
admin2selvars_list <- strsplit(admin2selvars_list, " +")[[1]]

admin1selvars_list <- readLines("data-clean/model_selection/selected_variables_admin1.txt")
admin1selvars_list <- strsplit(admin1selvars_list, " +")[[1]]


### put together the fay herriot level model
divisions_dt <- unique(grid_dt[, c("admin1Pcod", "admin1Name", "admin2Pcod", "admin2Name",
                                   "admin3Pcod", "admin3Name", "admin4Pcod", "admin4Name")])

#### remove the welf_int variables from the candidate variables
admin2_candvars <- colnames(admin2_dt)[!grepl("_welf_int",
                                              colnames(admin2_dt))]
admin2_candvars <- admin2_candvars[!grepl("admin2Pcod", admin2_candvars)]

### compare variance, design effect and effective sample size
geosurvey_dt <-
  geosurvey_dt %>%
  merge(geocodes_dt[, c("hhid", "enum_area")])

geosurvey_dt[, poor := ifelse(welfare < pline_int_215, 1, 0)]

geosvy_obj <- survey::svydesign(id = ~enum_area,
                                weights = ~hhweight,
                                survey.lonely.psu = "adjust",
                                data = geosurvey_dt)

areapop_dt <- svyby(~poor,
                    ~admin1Pcod,
                    geosvy_obj,
                    svymean,
                    deff = TRUE)

areapop_dt <- as.data.table(areapop_dt)

add_dt <- geosurvey_dt[, length(hhid), by = c("admin1Pcod", "admin2Pcod")] %>%
  setnames(old = "V1", new = "sample_size") ## compute size of sample in each target area

add_dt <- add_dt[!is.na(admin1Pcod),] ##drop missing observations (only 1)

areapop_dt <- add_dt[areapop_dt,
                     on = c("admin1Pcod")]

areapop_dt[, ess := sample_size / DEff.poor]

### re-estimate the actual poverty rates
pov_dt <-
  geosurvey_dt[, weighted.mean(x = poor,
                               w = hhweight,
                               na.rm = TRUE),
               by = c("admin1Pcod", "admin2Pcod")] %>%
  setnames(old = "V1", new = "Direct")

areapop_dt <- pov_dt[areapop_dt, on = c("admin1Pcod", "admin2Pcod")]

commune_dt <- areapop_dt[admin2_dt, on = "admin2Pcod"]


### check poverty rate
geosurvey_dt %>%
  summarise(weighted.mean(x = poor,
                          w = hhweight,
                          na.rm = TRUE))

### add population to commune level data
communepop_dt <-
  grid_dt[, sum(wpop_population, na.rm = TRUE),
          by = admin2Pcod] %>%
  setnames(old = "V1", new = "population")

commune_dt <- communepop_dt[commune_dt, on = "admin2Pcod"]

add_dt <- unique(grid_dt[, c(colnames(grid_dt)[grepl("BDI",
                                                     colnames(grid_dt))],
                             "admin2Pcod"),
                         with = FALSE])

commune_dt <- add_dt[commune_dt, on = "admin2Pcod"]

admin2_candvars <- c(admin2_candvars,
                     colnames(add_dt)[grepl("BDI",
                                           colnames(add_dt))])

haven::write_dta(commune_dt[, c("Direct",
                                admin2_candvars,
                                "population",
                                "admin1Pcod",
                                "admin2Pcod"), with = F],
                 "data-clean/commune_upddata.dta")

### next we perform model selection
selvars_list <- countrymodel_select(dt = commune_dt,
                                    xvars = admin2_candvars,
                                    y = "Direct")

### compute the fh model
domsize_dt <-
  geosurvey_dt[, sum(hhweight, na.rm = TRUE), by = admin2Pcod] %>%
  setnames(old = "V1", new = "domsize") %>%
  filter(!is.na(admin2Pcod))

saepop_dt <-
  sae::direct(y = poor,
              dom = admin2Pcod,
              sweight = hhweight,
              domsize = domsize_dt,
              data = geosurvey_dt) %>%
  mutate(var = SD ^ 2)

setnames(saepop_dt, "Domain", "admin2Pcod")

commune_dt$Direct <- NULL
saepop_dt <- as.data.table(saepop_dt)
commune_dt <- saepop_dt[commune_dt, on = "admin2Pcod"]



combine_dt <- povmap::combine_data(pop_data = commune_dt[, c(selvars_list,
                                                             "admin2Pcod"),
                                                         with = FALSE],
                                   pop_domains = "admin2Pcod",
                                   smp_data = commune_dt[!is.na(Direct),
                                                         c("admin2Pcod",
                                                           "Direct",
                                                           "var",
                                                           "ess"),
                                                         with = FALSE],
                                   smp_domains = "admin2Pcod")

fhmodel_not <-
  povmap::fh(fixed = as.formula(paste("Direct ~ ", paste(selvars_list, collapse= "+"))),
             vardir = "var",
             combined_data = combine_dt,
             domains = "admin2Pcod",
             method = "ml",
             MSE = TRUE,
             mse_type = "analytical")

fhmodel_log <-
  povmap::fh(fixed = as.formula(paste("Direct ~ ", paste(selvars_list, collapse= "+"))),
             vardir = "var",
             combined_data = combine_dt,
             domains = "admin2Pcod",
             method = "ml",
             MSE = TRUE,
             mse_type = "analytical",
             transformation = "log",
             backtransformation = "bc_sm")

fhmodel_arcsin <-
  povmap::fh(fixed = as.formula(paste("Direct ~ ", paste(selvars_list, collapse= "+"))),
             vardir = "var",
             combined_data = combine_dt,
             domains = "admin2Pcod",
             method = "ml",
             MSE = TRUE,
             mse_type = "boot",
             transformation = "arcsin",
             backtransformation = "bc",
             eff_smpsize = "ess")


result_dt <- as.data.table(fhmodel_arcsin$ind)

#### compare province level rates
fhpov_dt <-
  result_dt %>%
  merge(unique(commune_dt[, c("population",
                              "admin2Pcod",
                              "admin1Pcod")]),
        by.y = "admin2Pcod",
        by.x = "Domain") %>%
  group_by(admin1Pcod) %>%
  summarise(provFH = weighted.mean(x = FH,
                                   w = population,
                                   na.rm = TRUE)) %>%
  merge(geosurvey_dt %>%
          group_by(admin1Pcod) %>%
          summarise(provDirect = weighted.mean(x = poor,
                                               w = hhweight,
                                               na.rm = TRUE)),
        by = "admin1Pcod")


### save the fay herriot model results for arcsin
saveRDS(object = fhmodel_arcsin,
        "data-clean/fhmodel_arcsin_admin2.RDS")



#### run the povmap to estimate the poverty map at admin2, 3, 4 levels with log transformation
admin4_model <-
  povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin4selvars_list, collapse= "+"))),
              pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
                                                       c(admin4selvars_list,
                                                          "admin4Pcod",
                                                          "wpop_population"),
                                                       with = FALSE])),
              pop_domains = "admin4Pcod",
              smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
                                                            c("welfare",
                                                              admin4selvars_list,
                                                              "admin4Pcod",
                                                              "hhweight"),
                                                            with = FALSE])),
              smp_domains = "admin4Pcod",
              L = 100,
              B = 100,
              transformation = "log",
              threshold = 11111.76,
              weights = "hhweight",
              pop_weights = "wpop_population",
              cpus = 30,
              MSE = TRUE,
              na.rm = TRUE,
              rescale_weights = TRUE,
              Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin4_ydump_log.csv")

admin3_model <-
  povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
              pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
                                                       c(admin3selvars_list,
                                                         "admin3Pcod",
                                                         "wpop_population"),
                                                       with = FALSE])),
              pop_domains = "admin3Pcod",
              smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
                                                            c("welfare",
                                                              admin3selvars_list,
                                                              "admin3Pcod",
                                                              "hhweight"),
                                                            with = FALSE])),
              smp_domains = "admin3Pcod",
              L = 100,
              B = 100,
              transformation = "log",
              threshold = 11111.76,
              weights = "hhweight",
              pop_weights = "wpop_population",
              cpus = 30,
              MSE = TRUE,
              na.rm = TRUE,
              rescale_weights = TRUE,
              Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin3_ydump_log.csv")

admin2_model <-
  povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
              pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
                                                       c(admin2selvars_list,
                                                         "admin2Pcod",
                                                         "wpop_population"),
                                                       with = FALSE])),
              pop_domains = "admin2Pcod",
              smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
                                                            c("welfare",
                                                              admin2selvars_list,
                                                              "admin2Pcod",
                                                              "hhweight"),
                                                            with = FALSE])),
              smp_domains = "admin2Pcod",
              L = 100,
              B = 100,
              transformation = "log",
              threshold = 11111.76,
              weights = "hhweight",
              pop_weights = "wpop_population",
              cpus = 30,
              MSE = TRUE,
              na.rm = TRUE,
              rescale_weights = TRUE,
              Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin2_ydump_log.csv")

saveRDS(admin4_model, "data-clean/ebp_results/admin4_model.RDS")
saveRDS(admin3_model, "data-clean/ebp_results/admin3_model.RDS")
saveRDS(admin2_model, "data-clean/ebp_results/admin2_model.RDS")


#### run the povmap to estimate the poverty map at admin2, 3, 4 levels with log transformation
admin4_model <-
  povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
              pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
                                                       c(admin3selvars_list,
                                                         "admin4Pcod",
                                                         "wpop_population"),
                                                       with = FALSE])),
              pop_domains = "admin4Pcod",
              smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
                                                            c("welfare",
                                                              admin3selvars_list,
                                                              "admin4Pcod",
                                                              "hhweight"),
                                                            with = FALSE])),
              smp_domains = "admin4Pcod",
              L = 100,
              B = 100,
              transformation = "log",
              threshold = 11111.76,
              weights = "hhweight",
              pop_weights = "wpop_population",
              cpus = 30,
              MSE = TRUE,
              na.rm = TRUE,
              rescale_weights = TRUE,
              Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin4_ydump_log.csv")

admin3_model <-
  povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
              pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
                                                       c(admin3selvars_list,
                                                         "admin3Pcod",
                                                         "wpop_population"),
                                                       with = FALSE])),
              pop_domains = "admin3Pcod",
              smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
                                                            c("welfare",
                                                              admin3selvars_list,
                                                              "admin3Pcod",
                                                              "hhweight"),
                                                            with = FALSE])),
              smp_domains = "admin3Pcod",
              L = 100,
              B = 100,
              transformation = "log",
              threshold = 11111.76,
              weights = "hhweight",
              pop_weights = "wpop_population",
              cpus = 30,
              MSE = TRUE,
              na.rm = TRUE,
              rescale_weights = TRUE,
              Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin3_ydump_log.csv")

admin2_model <-
  povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
              pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
                                                       c(admin2selvars_list,
                                                         "admin2Pcod",
                                                         "wpop_population"),
                                                       with = FALSE])),
              pop_domains = "admin2Pcod",
              smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
                                                            c("welfare",
                                                              admin2selvars_list,
                                                              "admin2Pcod",
                                                              "hhweight"),
                                                            with = FALSE])),
              smp_domains = "admin2Pcod",
              L = 100,
              B = 100,
              transformation = "log",
              threshold = 11111.76,
              weights = "hhweight",
              pop_weights = "wpop_population",
              cpus = 30,
              MSE = TRUE,
              na.rm = TRUE,
              rescale_weights = TRUE,
              Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin2_ydump_log.csv")

saveRDS(admin4_model, "data-clean/ebp_results/admin4_model.RDS")
saveRDS(admin3_model, "data-clean/ebp_results/admin3_model.RDS")
saveRDS(admin2_model, "data-clean/ebp_results/admin2_model.RDS")




# ### same thing with logshift transformation
# admin4_model_logshift <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin2selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin2selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log.shift",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin4_ydump_logshift.csv")
#
# admin3_model_logshift <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin2selvars_list,
#                                                          "admin3Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin3Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin2selvars_list,
#                                                               "admin3Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin3Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log.shift",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin3_ydump_logshift.csv")
#
# admin2_model_logshift <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin2selvars_list,
#                                                          "admin2Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin2selvars_list,
#                                                               "admin2Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log.shift",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin2_ydump_logshift.csv")
#
# saveRDS(admin4_model_logshift, "data-clean/ebp_results/admin4_model_logshift.RDS")
# saveRDS(admin3_model_logshift, "data-clean/ebp_results/admin3_model_logshift.RDS")
# saveRDS(admin2_model_logshift, "data-clean/ebp_results/admin2_model_logshift.RDS")
#
#
#
# #### models created from the variable selection on admin3 clustering
# admin4_model3 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin2_ydump3.csv")
#
# admin3_model3 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin3Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin3Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin3Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin3Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin3_ydump3.csv")
#
# admin2_model3 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin2Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin2Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin3_ydump3.csv")
#
# saveRDS(admin4_model3, "data-clean/ebp_results/admin4_model3.RDS")
# saveRDS(admin3_model3, "data-clean/ebp_results/admin3_model3.RDS")
# saveRDS(admin2_model3, "data-clean/ebp_results/admin2_model3.RDS")
#
#
# #### models created from the variable selection on admin4 clustering
#
# admin4_model4 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin4selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin4selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin4selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin4_ydump4.csv")
#
# admin3_model4 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin4selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin4selvars_list,
#                                                          "admin3Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin3Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin4selvars_list,
#                                                               "admin3Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin3Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin3_ydump4.csv")
#
# admin2_model4 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin4selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin4selvars_list,
#                                                          "admin2Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin4selvars_list,
#                                                               "admin2Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin2_ydump4.csv")
#
# saveRDS(admin4_model4, "data-clean/ebp_results/admin4_model4.RDS")
# saveRDS(admin3_model4, "data-clean/ebp_results/admin3_model4.RDS")
# saveRDS(admin2_model4, "data-clean/ebp_results/admin2_model4.RDS")
#
#
#
#
# #### models created from the variable selection on admin3 clustering using boxcox transformation
# admin4_model3_boxcox <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "box.cox",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin4_ydump3_boxcox.csv")
#
# admin3_model3_boxcox <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin3Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin3Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin3Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin3Pcod",
#               L = 100,
#               B = 100,
#               transformation = "box.cox",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin3_ydump3_boxcox.csv")
#
# admin2_model3_boxcox <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin2Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin2Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "box.cox",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin2_ydump3_boxcox.csv")
#
# saveRDS(admin4_model3_boxcox, "data-clean/ebp_results/admin4_model3_boxcox.RDS")
# saveRDS(admin3_model3_boxcox, "data-clean/ebp_results/admin3_model3_boxcox.RDS")
# saveRDS(admin2_model3_boxcox, "data-clean/ebp_results/admin2_model3_boxcox.RDS")
#
#
# #### models created from the variable selection on admin3 clustering using ordernorm transformation
# admin4_model3_ordernorm <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "ordernorm",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin4_ydump3_ordernorm.csv")
#
# admin3_model3_ordernorm <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin3Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin3Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin3Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin3Pcod",
#               L = 100,
#               B = 100,
#               transformation = "ordernorm",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin3_ydump3_ordernorm.csv")
#
# admin2_model3_ordernorm <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin2Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin2Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "ordernorm",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin2_ydump3_ordernorm.csv")
#
# saveRDS(admin4_model3_ordernorm, "data-clean/ebp_results/admin4_model3_ordernorm.RDS")
# saveRDS(admin3_model3_ordernorm, "data-clean/ebp_results/admin3_model3_ordernorm.RDS")
# saveRDS(admin2_model3_ordernorm, "data-clean/ebp_results/admin2_model3_ordernorm.RDS")
#
# ### run some more models for comparison
# admin2_model_logshift <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin2selvars_list,
#                                                          "admin2Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin2selvars_list,
#                                                               "admin2Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log.shift",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin2_ydump_logshift.csv")
#
# admin2_model_ordernorm <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin2selvars_list,
#                                                          "admin2Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin2selvars_list,
#                                                               "admin2Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "ordernorm",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin2_ydump_ordernorm.csv")
#
# admin2_model_boxcox <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin2selvars_list,
#                                                          "admin2Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin2selvars_list,
#                                                               "admin2Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "box.cox",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin2_ydump_boxcox.csv")
#
# #saveRDS(admin2_model_logshift, "data-clean/ebp_results/admin2_model_logshift.RDS")
# saveRDS(admin2_model_ordernorm, "data-clean/ebp_results/admin2_model_ordernorm.RDS")
# saveRDS(admin2_model_boxcox, "data-clean/ebp_results/admin2_model_boxcox.RDS")
#
#
# admin4_model4_logshift <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin4selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin4selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin4selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log.shift",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin4_ydump4_logshift.csv")
#
# admin4_model4_ordernorm <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin4selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin4selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin4selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "ordernorm",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin4_ydump4_ordernorm.csv")
#
# admin4_model4_boxcox <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin4selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin4selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin4selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "box.cox",
#               threshold = 11111.76,
#               weights = "hhweight",
#               weights_type = "nlme",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/admin4_ydump4_boxcox.csv")
#
# saveRDS(admin4_model4_logshift, "data-clean/ebp_results/admin4_model4_logshift.RDS")
# saveRDS(admin4_model4_ordernorm, "data-clean/ebp_results/admin4_model4_ordernorm.RDS")
# saveRDS(admin4_model4_boxcox, "data-clean/ebp_results/admin4_model4_boxcox.RDS")
#
#
# #### models with area dummies since we have forgotten them so far
# admin4_model1 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin1selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin1selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin1selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin4_ydump1.csv")
#
# admin3_model1 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin1selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin1selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin3Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin1selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin3Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin3_ydump1.csv")
#
# admin2_model1 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin1selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin1selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin1selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin2_ydump1.csv")
#
# saveRDS(admin4_model1, "data-clean/ebp_results/adm1_included/admin4_model1.RDS")
# saveRDS(admin3_model1, "data-clean/ebp_results/adm1_included/admin3_model1.RDS")
# saveRDS(admin2_model1, "data-clean/ebp_results/adm1_included/admin2_model1.RDS")
#
#
# ### compute models from variables selected via admin2 clustering in model selection
# admin4_model2 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin2selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin2selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin4_ydump2.csv")
#
# admin3_model2 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin2selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin3Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin2selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin3Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin3_ydump1.csv")
#
# admin2_model2 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin2selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin2selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin2selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin2_ydump1.csv")
#
# saveRDS(admin4_model2, "data-clean/ebp_results/adm1_included/admin4_model2.RDS")
# saveRDS(admin3_model2, "data-clean/ebp_results/adm1_included/admin3_model2.RDS")
# saveRDS(admin2_model2, "data-clean/ebp_results/adm1_included/admin2_model2.RDS")
#
# ### compute models from variables selected via admin3 clustering in model selection
# admin4_model3 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin4_ydump3.csv")
#
# admin3_model3 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin3Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin3Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin3_ydump3.csv")
#
# admin2_model3 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin3selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin3selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin3selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin2_ydump3.csv")
#
# saveRDS(admin4_model3, "data-clean/ebp_results/adm1_included/admin4_model3.RDS")
# saveRDS(admin3_model3, "data-clean/ebp_results/adm1_included/admin3_model3.RDS")
# saveRDS(admin2_model3, "data-clean/ebp_results/adm1_included/admin2_model3.RDS")
#
#
# ### compute models from variables selected via admin3 clustering in model selection
# admin4_model4 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin4selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin4selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin4Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin4selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin4Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin4_ydump4.csv")
#
# admin3_model4 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin4selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin4selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin3Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin4selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin3Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin3_ydump4.csv")
#
# admin2_model4 <-
#   povmap::ebp(fixed = as.formula(paste("welfare ~ ", paste(admin4selvars_list, collapse= "+"))),
#               pop_data = as.data.frame(na.omit(grid_dt[wpop_population > 0,
#                                                        c(admin4selvars_list,
#                                                          "admin4Pcod",
#                                                          "wpop_population"),
#                                                        with = FALSE])),
#               pop_domains = "admin2Pcod",
#               smp_data = as.data.frame(na.omit(geosurvey_dt[!is.na(admin4Pcod),
#                                                             c("welfare",
#                                                               admin4selvars_list,
#                                                               "admin4Pcod",
#                                                               "hhweight"),
#                                                             with = FALSE])),
#               smp_domains = "admin2Pcod",
#               L = 100,
#               B = 100,
#               transformation = "log",
#               threshold = 11111.76,
#               weights = "hhweight",
#               pop_weights = "wpop_population",
#               cpus = 30,
#               MSE = TRUE,
#               na.rm = TRUE,
#               rescale_weights = TRUE,
#               Ydump = "//esapov/esapov/BDI/GEO/Population/ydump/adm1_included/admin2_ydump4.csv")
#
# saveRDS(admin4_model4, "data-clean/ebp_results/adm1_included/admin4_model4.RDS")
# saveRDS(admin3_model4, "data-clean/ebp_results/adm1_included/admin3_model4.RDS")
# saveRDS(admin2_model4, "data-clean/ebp_results/adm1_included/admin2_model4.RDS")


save.image("//esapov/esapov/BDI/GEO/Population/unitmodel_images.RData")
save.image("data-raw/unitmodel_images.RData")
